readded some user checks, due to a refactor

+++ src/main/kotlin/com/rentmycar/services/TimeSlotService.kt
@@ -2,6 +2,7 @@ package com.rentmycar.services
 
 import com.rentmycar.entities.Car
 import com.rentmycar.dtos.TimeslotDTO
+import com.rentmycar.entities.User
 import com.rentmycar.entities.toDTO
 import com.rentmycar.repositories.TimeSlotRepository
 import com.rentmycar.services.exceptions.NotAllowedException
@@ -15,7 +16,9 @@ import kotlinx.datetime.toLocalDateTime
 class TimeSlotService {
     private val timeSlotRepository = TimeSlotRepository()
 
-    fun createTimeSlot(carId: Int, timeSlotRange: OpenEndRange<LocalDateTime>) {
+    fun createTimeSlot(user: User, carId: Int, timeSlotRange: OpenEndRange<LocalDateTime>) {
+        CarService.ensureUserIsCarOwner(user, carId)
+
         val car = CarService.getBusinessObject(carId).getCar()
 
         if (timeSlotRepository.getOverlappingTimeSlots(car, timeSlotRange)
@@ -34,6 +37,7 @@ class TimeSlotService {
     fun getTimeSlots(car: Car) = timeSlotRepository.getTimeSlots(car)
 
     fun updateTimeSlot(
+        user: User,
         timeSlotId: Int,
         startDateTime: LocalDateTime? = null,
         endDateTime: LocalDateTime? = null,
@@ -45,6 +49,8 @@ class TimeSlotService {
         val updatedEndDateTime = endDateTime ?: timeSlotDTO.availableUntil
 
         val updatedTimeSlotRange = updatedStartDateTime.rangeUntil(updatedEndDateTime)
+
+        CarService.ensureUserIsCarOwner(user, timeSlotDTO.carId)
         val car = CarService.getBusinessObject(timeSlotDTO.carId).getCar()
 
         if (timeSlotRepository.getOverlappingTimeSlots(car, updatedTimeSlotRange)
@@ -56,9 +62,11 @@ class TimeSlotService {
         timeSlotRepository.updateTimeSlot(timeSlot, updatedTimeSlotRange)
     }
 
-    fun deleteTimeSlot(id: Int) {
+    fun deleteTimeSlot(user: User, id: Int) {
         val timeSlot = getTimeSlot(id)
         if (isFutureTimeSlot(timeSlot.toDTO())) throw throw NotAllowedException("cannot delete an active or past timeslot")
+
+        CarService.ensureUserIsCarOwner(user, timeSlot.toDTO().carId)
         CarService.getBusinessObject(timeSlot.toDTO().carId)
 
         timeSlotRepository.deleteTimeSlot(timeSlot)
